% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{add_na_df}
\alias{add_na_df}
\title{Replace NULLs in list with empty data.frames}
\usage{
add_na_df(list_of_df)
}
\arguments{
\item{list_of_df}{A list of data.frame.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A list.
}\if{html}{\out{</div>}}
}
\description{
Replace NULLs in list with empty data.frames. The data.frames in
the list have the same columns and data types.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
