% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{BRIEFCASE.SKIP}
\alias{BRIEFCASE.SKIP}
\title{Number of lines to skip from a briefcase metadata file.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
BRIEFCASE.SKIP
}
\description{
Number of lines to skip from a briefcase metadata file.
}
\keyword{datasets}
