% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{BRIEFCASE.VALID.FLAGS}
\alias{BRIEFCASE.VALID.FLAGS}
\title{Flags for identifying valid observations in briefcase data.}
\format{
An object of class \code{character} of length 3.
}
\usage{
BRIEFCASE.VALID.FLAGS
}
\description{
Flags for identifying valid observations in briefcase data.
}
\keyword{datasets}
