% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{build_date}
\alias{build_date}
\title{Build a date}
\usage{
build_date(year, month, day)
}
\arguments{
\item{year, month, day}{vectors (of the same length) which
together constitute a date.}
}
\value{
a date vector.
}
\description{
Build a date from its parts.
}
