% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{build_grid}
\alias{build_grid}
\title{Build a grid}
\usage{
build_grid(
  origin_lon,
  origin_lat,
  min_lon,
  max_lon,
  min_lat,
  max_lat,
  grid_resolution,
  crs,
  grid_id = "grid_id"
)
}
\arguments{
\item{origin_lon, origin_lat}{a numeric(1). Coordinate of origin.}

\item{min_lon, max_lon}{a numeric(1). Grid's mininum and maximum longitude
value.}

\item{min_lat, max_lat}{a numeric(1). Grid's mininum and maximum latitudes
values.}

\item{grid_resolution}{a numeric. Grid's resolution.}

\item{crs}{a numeric(1). Coordinate reference system (EPSG).}

\item{grid_id}{a character(1). The name for the columns with the identifier
for each cell in the grid.}
}
\value{
an sf object (polygon).
}
\description{
Build a grid that includes the given origin as a vertex.
}
