% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_area.R
\name{compute_frequency_grid}
\alias{compute_frequency_grid}
\title{Aggregate trajectories' vertices using a grid}
\usage{
compute_frequency_grid(
  traj_df,
  grid_sf,
  clon = "lon",
  clat = "lat",
  cheight = "height",
  grid_id = "grid_id",
  crs = sf::st_crs(grid_sf),
  densify_n = 1
)
}
\arguments{
\item{traj_df}{a data frame. Data representing a trajectory.}

\item{grid_sf}{an sf object (polygons). The grid used to aggregate
trajectories.}

\item{clon, clat, cheight}{a character(1). Names of the longitude, latitude,
and height columns in \code{traj_df}.}

\item{grid_id}{a character(1). Name of the column in \code{grid_sf} which
identifies each cell in the grid.}

\item{crs}{a numeric. EPSG code used for the trajectory data in \code{traj_df}.}

\item{densify_n}{an integer. Increment the number of vertices in each of
the segments in the given trajectory (see 'smoothr::densify').}
}
\value{
an sf object. The given grid_sf object with additional attributes
("freq").
}
\description{
Compute the number of trajectories' vertices for each cell in the grid.
}
\seealso{
\link{build_grid} for building grids.
}
