% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\name{computeTrajTime}
\alias{computeTrajTime}
\title{Compute the trajectory time}
\usage{
computeTrajTime(file.vec, line.vec, cnames)
}
\arguments{
\item{file.vec}{A vector of character. The paths to the input files}

\item{line.vec}{A vector. Ids of rows in each file in file.vec}

\item{cnames}{A character. Column names of hysplit files}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    A list of difftime (in days)
}\if{html}{\out{</div>}}
}
\description{
Compute the trajectory time
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
