% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\docType{class}
\name{cqmaTools}
\alias{cqmaTools}
\alias{cqmaTools-class}
\title{The CQMATOOLS class}
\usage{
cqmaTools()
}
\description{
Use this class for representing a CQMATOOLS object

Creates a cqmaTools object
}
\note{
No notes
}
\author{
Alber Sanchez
}

