% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{filename2profile}
\alias{filename2profile}
\title{Build a profile name from trajectory paths}
\usage{
filename2profile(files, traj_pat = "^[A-Z]{3}_[0-9]{4}_[0-9]{2}_[0-9]{2}.")
}
\arguments{
\item{files}{a character. Path to trajectory files.}

\item{traj_pat}{a character. Regular expression pattern that the
trajectories' filenames must match.}
}
\value{
a character.
}
\description{
Build a profile name using the names of the given files. The profile names
are made of the first 4 elements of the trajectories' filenames: site, year,
month, and day.
}
