% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.R
\name{files2df}
\alias{files2df}
\title{Read files into data frames}
\usage{
files2df(files, header, skip, cnames)
}
\arguments{
\item{files}{a character. Paths to files.}

\item{header}{a logical(1). Do the files have a header row?}

\item{skip}{a numeric(1). Number of lines to skip from each file.}

\item{cnames}{a character. Column names for the data in files.}
}
\value{
a list of data frames.
}
\description{
Read the given files into data frames (one per file).
}
