% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.R
\name{filter_data_frames}
\alias{filter_data_frames}
\title{Filter data frames}
\usage{
filter_data_frames(x, cname, min, max)
}
\arguments{
\item{x}{either a data frame or a list of data frames.}

\item{cname}{a character(1). A column name in each of the given data frames.}

\item{min, max}{numeric(1). Maximum and mininum values.}
}
\value{
a data frame or a list of data frames or NA (see details).
}
\description{
Remove data frames from a list when their rows don't fall inside the given
numeric range for the given column.
}
\details{
When x is a single data frame and it doesn't meet the filter, this function
returns NA.
}
