% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{floorconcentration}
\alias{floorconcentration}
\title{Compute the concentration on the ground}
\usage{
floorconcentration(dif_obs_bkg, hfloor, temp, height, molair)
}
\arguments{
\item{dif_obs_bkg}{A numeric. The difference between the observed conetration and the background concentration}

\item{hfloor}{A numeric. The hieght of the floor}

\item{temp}{A numeric. The tempearture of the first observation above the floor}

\item{height}{A numeric. The height of the first observaton above the floor}

\item{molair}{A numeric. The molarity of the air}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        A numeric. The concentration on the ground
}\if{html}{\out{</div>}}
}
\description{
Compute the concentration on the ground
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
