% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{format_traj_names}
\alias{format_traj_names}
\title{Format trajectories' names}
\usage{
format_traj_names(traj_names, pattern = TRAJECTORY.FILENAME.PATTERN)
}
\arguments{
\item{traj_names}{a character. Names of trajectories.}

\item{pattern}{a character(1). Pattern of valid trajectory names.}
}
\value{
a character.
}
\description{
Put the height before the hour in given trajectories' names. Also, ensure
the height has the same number of digits.
}
