% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggplot2.R
\name{get_map_plot}
\alias{get_map_plot}
\title{Build a map trajectories}
\usage{
get_map_plot(data_tb, cid, clon, clat, range_lon = NA, range_lat = NA)
}
\arguments{
\item{data_tb}{A tibble with backtrajectory data.}

\item{cid}{A character(1). Name of the column that identifies each
backtrajectory}

\item{clon, clat}{A character(1). Names of the columns with the longitude and
latitude.}

\item{range_lon, range_lat}{a double(2). Minimum and maximum longitude and
latitude of the map.}
}
\value{
A ggplot object.
}
\description{
Build a map of backtrajectories.
}
