% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.R
\name{get_min_max}
\alias{get_min_max}
\title{Get minimum and maximum}
\usage{
get_min_max(.data, suffix_min = "min", suffix_max = "max")
}
\arguments{
\item{.data}{a data frame.}

\item{suffix_min, suffix_max}{a charter. Suffixes for the column names in
resulting data frame.}
}
\value{
a data.frame.
}
\description{
Get the mininum and maximum values of the numeric columns in the given data
frame.
}
