% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{get_raster_range}
\alias{get_raster_range}
\title{Get the rage of raster values}
\usage{
get_raster_range(r)
}
\arguments{
\item{r}{a 'terra' raster or a list of them.}
}
\value{
a numeric.
}
\description{
Get the mininum and maximum value in the given rasters.
}
