% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{get_traj_extent}
\alias{get_traj_extent}
\title{Get the trajectories' extent}
\usage{
get_traj_extent(.data, ext_cnames = c("lat", "lon", "height", "date"))
}
\arguments{
\item{.data}{a data frame of a list of data frame.}

\item{ext_cnames}{a characcter. Column names representing varibles on which
the extent is computed.}
}
\value{
a data frame or a list of data frames.
}
\description{
Get the extent of the given trajectories.
}
