% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{grid_helper}
\alias{grid_helper}
\title{Build a vector that passes through the origin}
\usage{
grid_helper(o, min, max, res)
}
\arguments{
\item{o}{a numeric(1). Origin.}

\item{min}{a numeric(1). Mininum value.}

\item{max}{a numeric(1). Maximum value.}

\item{res}{a numeric(1). Grid resultuion.}
}
\value{
A numeric.
}
\description{
Build a vector from the minimum to the maximum value ensuring that the
given origin is a value in the returned vector.
}
