% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{HYSPLIT.COLNAMES}
\alias{HYSPLIT.COLNAMES}
\title{The name of the columns in the trajectory files produced by the Hysplit software.}
\format{
An object of class \code{character} of length 13.
}
\usage{
HYSPLIT.COLNAMES
}
\description{
The name of the columns in the trajectory files produced by the Hysplit software.
}
\keyword{datasets}
