% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{in_interval}
\alias{in_interval}
\title{Test in which interval are the given values}
\usage{
in_interval(val, vec)
}
\arguments{
\item{val}{a numeric vector of the values to test.}

\item{vec}{a numeric vector with the limits of the consecutive intervals.
Its length must be greater than 1.}
}
\value{
a boolean matrix. The columns represent the interval's limits (vec)
and the rows the values (val). The number of intervals is the number of
element in vec minus one
}
\description{
Test in which interval are the given values
}
