% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\name{intersectTraj.intersect}
\alias{intersectTraj.intersect}
\title{Invert the parameters of rgeos::gIntersection}
\usage{
intersectTraj.intersect(g2, g1, byid)
}
\arguments{
\item{g2}{A SP object}

\item{g1}{A SP object}

\item{byid}{A logical vector to pass to rgeos::gIntersection}
}
\value{
The intersection resutl as SP objects
}
\description{
Invert the parameters of rgeos::gIntersection
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
