% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{moving_window}
\alias{moving_window}
\title{Apply a function to a vector using a moving window}
\usage{
moving_window(x, f, hsize, ...)
}
\arguments{
\item{x}{a numeric vector.}

\item{f}{a function that receives a numeric and returns a numeric(1).}

\item{hsize}{an integer(1). Number of elements to take before and after
to build the moving window.}

\item{...}{other arguments passed to f.}
}
\value{
A numeric the same size as the input vector.
}
\description{
Apply the given function to the given numeric vector using a moving window.
The window size twice (plus one) the given hsize.
}
