% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_area.R
\name{plot_influence_area}
\alias{plot_influence_area}
\title{Plot influence area}
\usage{
plot_influence_area(
  r,
  r_range = range(r[]),
  r_col = terra::map.pal("viridis", 100),
  x_range = c(-180, 180),
  y_range = c(-90, 90),
  add_countries = TRUE,
  ctr_color = "black",
  ctr_lwd = 2,
  add_states = TRUE,
  stt_color = "gray",
  stt_lwd = 1,
  plot_title = "",
  save_plots = FALSE,
  plot_width = 480,
  plot_height = 480
)
}
\arguments{
\item{r}{a raster. The area of influence.}

\item{r_range}{a numeric(2). The range of values used during plot.}

\item{r_col}{a color palette for mapping. See \link[terra:mappal]{terra::map.pal}.}

\item{x_range, y_range}{a numeric(2). The plot range in x and y.}

\item{add_countries}{a logical(1). Should countries' borderds be added?}

\item{ctr_color}{a character(1). Color name used to plot the countries.}

\item{ctr_lwd}{a numeric(1). Line width used to plot the countries.}

\item{add_states}{a logical(1). Should we add the states to the plot?}

\item{stt_color}{a character(1). Color name for the Brazilian states.}

\item{stt_lwd}{a numeric(1). Line width for the Brazilian states.}

\item{plot_title}{a character(1). Title of the plot.}

\item{save_plots}{either a logical(1) or a character(1). Should the plot be
saved to files? If so, a path to a directory should be provided,
otherwise, the current directory is used. The file name is taken from
\code{plot_title}.}

\item{plot_width, plot_height}{a numeric(1). Size of the plot.}
}
\value{
a character. The path to the files created.
}
\description{
Plot the given influence area (a raster).
}
