% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggplot2.R
\name{plot_trajectories}
\alias{plot_trajectories}
\title{Plot a set of trajectories}
\usage{
plot_trajectories(
  files,
  cnames = HYSPLIT.COLNAMES,
  header = FALSE,
  skip = 7,
  xlim = c(-80, -30),
  ylim = c(-40, 10)
)
}
\arguments{
\item{files}{a character. Paths to trajectory files.}

\item{cnames}{a character. Names for the columns in the given files.}

\item{header}{a logical(1). Do files have headers?}

\item{skip}{an integer(1). Number of lines to skip from each file.}

\item{xlim, ylim}{a numeric(2). Longitude and latitude ranges of the
resulting plot (map).}
}
\value{
A ggplot object
}
\description{
Plot the trajectories in the input files.
}
