% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\name{plotTrajYear}
\alias{plotTrajYear}
\title{Plot the trajectories grouped by year}
\usage{
plotTrajYear(
  file.vec,
  path.out,
  device,
  map.xlim,
  map.ylim,
  map.height,
  map.width,
  stations.df,
  plot2file
)
}
\arguments{
\item{file.vec}{A character vector. The paths to  trajectory files}

\item{path.out}{A character. The path to the folder for storing the resulting files}

\item{device}{A character. Image format, i.e. PNG}

\item{map.xlim}{A numeric vector. Map's min & max longitude}

\item{map.ylim}{A numeric vector. Map's min & max latitude}

\item{map.height}{A numeirc. Map image size}

\item{map.width}{A numeirc. Map image size}

\item{stations.df}{A data.frame with metereological station data. It must contain at least the columns c("name", "lon", "lat")}

\item{plot2file}{A logical. Should the plot be stored as a file}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{        A list. The path to the created files
}\if{html}{\out{</div>}}
}
\description{
Plot the trajectories grouped by year
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
