% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RAWDATA.COLNAMES}
\alias{RAWDATA.COLNAMES}
\title{The name of the columns in the file resulting from analyzing gas samples.}
\format{
An object of class \code{character} of length 23.
}
\usage{
RAWDATA.COLNAMES
}
\description{
The name of the columns in the file resulting from analyzing gas samples.
}
\keyword{datasets}
