% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RAWDATA.SKIP}
\alias{RAWDATA.SKIP}
\title{Number of lines to skip from a rawdata file.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
RAWDATA.SKIP
}
\description{
Number of lines to skip from a rawdata file.
}
\keyword{datasets}
