% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{RAWDATA.VALID.FLAGS}
\alias{RAWDATA.VALID.FLAGS}
\title{Flags for identifying valid observations in raw data.}
\format{
An object of class \code{character} of length 3.
}
\usage{
RAWDATA.VALID.FLAGS
}
\description{
Flags for identifying valid observations in raw data.
}
\keyword{datasets}
