% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/briefcases.R
\name{read_briefcase_file}
\alias{read_briefcase_file}
\title{Read briefcase data from a file}
\usage{
read_briefcase_file(
  file_path,
  cnames = names(BRIEFCASE.COLNAMES),
  ctypes = BRIEFCASE.COLNAMES,
  skip = BRIEFCASE.SKIP
)
}
\arguments{
\item{file_path}{A character. Path to a data file.}

\item{cnames}{A character. Column names.}

\item{ctypes}{A character. Column types.}

\item{skip}{An integer. Number of lines to skip when reading the file.}
}
\value{
a data frame.
}
\description{
Read a file with pre-processed metadata from the briefcases.
}
