% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata.R
\name{read_rawdata_file}
\alias{read_rawdata_file}
\title{Read raw data from a file}
\usage{
read_rawdata_file(
  file_path,
  cnames = names(RAWDATA.COLNAMES),
  ctypes = RAWDATA.COLNAMES,
  skip = RAWDATA.SKIP
)
}
\arguments{
\item{file_path}{A character. Path to a data file.}

\item{cnames}{A character. Column names.}

\item{ctypes}{A character. Column types.}

\item{skip}{An integer. Number of lines to skip when reading the file.}
}
\value{
a data frame.
}
\description{
Read a file with data from the results of LaGEE analysis of GHGs.
}
