% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{read_station_file}
\alias{read_station_file}
\title{Read metereological data from a station}
\usage{
read_station_file(file_path)
}
\arguments{
\item{file_path}{A character. Path to a data file.}
}
\value{
a data.frame with 2 columns: date and value.
}
\description{
Read a file with metereological data.
}
