% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{read_trajectory_file}
\alias{read_trajectory_file}
\title{Read trajectory data from a file}
\usage{
read_trajectory_file(
  file_path,
  cnames = names(TRAJECTORY.COLNAMES),
  ctypes = TRAJECTORY.COLNAMES,
  skip = TRAJECTORY.SKIP
)
}
\arguments{
\item{file_path}{A character. Path to a data file.}

\item{cnames}{A character. Column names.}

\item{ctypes}{A character. Column types.}

\item{skip}{An integer. Number of lines to skip when reading the
trajectory file.}
}
\value{
a data frame.
}
\description{
Read a file with data from a backtrajectory.
}
