% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\docType{methods}
\name{runsimulation}
\alias{runsimulation}
\alias{runsimulation,character,numeric,character,character,character,character-method}
\alias{runsimulation-generic}
\title{Run HYSPLIT simulations for the traj_spec files in the given path}
\usage{
runsimulation(path.in, backTrajTime, hysplit.exec.path, hysplit.work.path,
  path.out, timezone)


  \S4method{runsimulation}{character,numeric,character,character,character,character}(path.in,
  backTrajTime, hysplit.exec.path, hysplit.work.path, path.out, timezone)
}
\arguments{
\item{path.in}{A character representing a folder with traj_spec files}

\item{backTrajTime}{A numeric. Time to modify the hysplit file search. For example, 10 days to the past is (10 * 24 * 3600) * (-1)}

\item{hysplit.exec.path}{A path to the HYSPLIT's exec folder.}

\item{hysplit.work.path}{A path to the HYSPLIT's working folder.}

\item{path.out}{A character. The path to store the resulting files. Usually HYSPLIT's simulations folder.}

\item{timezone}{A character. The time zone used for date computations.i.e. GMT}
}
\value{
A character list of the created files
}
\description{
Run HYSPLIT simulations for the traj_spec files in the given path
}

