% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_area.R
\name{smooth_neigh_na}
\alias{smooth_neigh_na}
\title{Smooth raster using NA neighbors}
\usage{
smooth_neigh_na(r, w = 3, threshold = 7)
}
\arguments{
\item{r}{A raster (terra).}

\item{w}{An integer. The window size.}

\item{threshold}{An integer.}
}
\value{
A raster (terra).
}
\description{
Transform the cells into NA depending on the number of NA
neighbors.
}
\author{
Alber Sánchez, \email{alber.ipia@inpe.br}
}
