% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{splines_interpolation}
\alias{splines_interpolation}
\title{Interpolate using splines}
\usage{
splines_interpolation(x, y, min_obs = 4)
}
\arguments{
\item{x}{a numeric. The x variable.}

\item{y}{a numeric. The y variable.}

\item{min_obs}{an integer(1). Mininum number of valid observations to fit a
model.}
}
\value{
A numeric with the same lenght as the input data.
}
\description{
Fit splines to the data and interpolate the missing observations.
}
