% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\name{splitRawdata}
\alias{splitRawdata}
\title{Split a raw data file.}
\usage{
splitRawdata(
  file.in,
  path.out,
  colname,
  keepFlags,
  cnames,
  keepCols,
  cnamesTest
)
}
\arguments{
\item{file.in}{A character.Path to a raw data file}

\item{path.out}{A character. Path to the folder for storing the resulting files}

\item{colname}{A character. The name of a column where the flags are located}

\item{keepFlags}{A character vector. Flags to keep in the raw data. i.e. c("...", "..>")}

\item{cnames}{A character vector. The name of the columns of the raw data file}

\item{keepCols}{A character vector. The column names of the raw data file to keep after filtering}

\item{cnamesTest}{A character vector. The column names used for testing duplicated rows in the filtered data}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{      A list of three. A character with the name of the new file, a data.frame of duplicated or inconsistent rows, and a data.frame of coordinates which were changed because they have the wrong sign
}\if{html}{\out{</div>}}
}
\description{
Split a single raw data file. The output files are prefixed with
"traj_specs_".
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
