% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states_sf}
\alias{states_sf}
\title{State polygons}
\format{
An sf object with 4596 rows and 2 columns:
name: State name.
geometry: Simple feature geometry list column sfc.
}
\source{
\url{https://www.naturalearthdata.com}
}
\usage{
data(states_sf)
}
\description{
Administrative level 1 polygons obtained from Natural Earth. These polygons
are used as plot background.
}
\keyword{datasets}
