% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TRAJECTORY.COLNAMES}
\alias{TRAJECTORY.COLNAMES}
\title{Names of the columns in the files with trajectory data.}
\format{
An object of class \code{character} of length 13.
}
\usage{
TRAJECTORY.COLNAMES
}
\description{
These names and descriptions were taken from the Hysplit users' guide (April
2025); the End Point file format, record loop 6 through the number of hours
in the simulation.
}
\details{
the column names and their meaning.
\itemize{
\item parameter traj_number Trajectory number.
\item parameter grid_number Metereological grid number or antecedent trajectory number.
\item parameter year Year of the point.
\item parameter month Month of the point.
\item parameter day Day of the point.
\item parameter hour Hour of the point.
\item parameter minute Minute of the point.
\item parameter forecast Forecast hour at point
\item parameter traj_age Age of the trajectory in hours.
\item parameter latitude Position latitude.
\item parameter longitude Position longitude.
\item parameter height Position height in meters above ground.
\item parameter pressure n diagnostic output variables; 1st to be output is always pressure.
}
}
\keyword{datasets}
