% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TRAJECTORY.FILENAME.PATTERN}
\alias{TRAJECTORY.FILENAME.PATTERN}
\title{Pattern of trajectories' filenames.}
\format{
An object of class \code{character} of length 1.
}
\usage{
TRAJECTORY.FILENAME.PATTERN
}
\description{
Pattern used for matching file names corresponding to
trajectory files. This is closely relate to  TRAJECTORY.METADATA.
}
\keyword{datasets}
