% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TRAJECTORY.METADATA}
\alias{TRAJECTORY.METADATA}
\title{Metadata contained in trajectories' file names.}
\format{
An object of class \code{character} of length 6.
}
\usage{
TRAJECTORY.METADATA
}
\description{
Description of metadata in the names of the trajectories' files. This is
closely related to TRAJECTORY.FILENAME.PATTERN.
}
\keyword{datasets}
