% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TRAJECTORY.SKIP}
\alias{TRAJECTORY.SKIP}
\title{Number of lines to skip from a trajectory file.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
TRAJECTORY.SKIP
}
\description{
Number of lines to skip from a trajectory file.
}
\keyword{datasets}
