% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\name{trajinbound}
\alias{trajinbound}
\title{Are trajectories inside boundaries?}
\usage{
trajinbound(traj.intersections, minx, maxx, miny, maxy)
}
\arguments{
\item{traj.intersections}{A list of two obejcts. A character vector and a list of data.frames. The data.frames are made of the rows in the matching trajectory file which first crossed to the sea}

\item{minx}{A numeric. The minimum X}

\item{maxx}{A numeric. The maximum Y}

\item{miny}{A numeric. The minimum Y}

\item{maxy}{A numeric. The maximum X}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{              A data frame
}\if{html}{\out{</div>}}
}
\description{
Check if the trajectory intersection is in the given boundary
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
