% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\name{trajOutInterpolation}
\alias{trajOutInterpolation}
\title{Are trajectories inside the metereological station coverage?}
\usage{
trajOutInterpolation(traj.intersections, stations.df)
}
\arguments{
\item{traj.intersections}{A list made of a character vector and a list. The character vector is the path to each trajectory file while the list contains the first row in the trajectory file which lies over the sea}

\item{stations.df}{A data.frame of metereological station data. It must have at least 4 columns: name, longitude, latitude, and file c("name", "lon", "lat", "file")}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{              A data.frame with one row for each file and 2 columns: The trajectories' path and a boolean indicating if meet the test
}\if{html}{\out{</div>}}
}
\description{
Check if the trajectories' intersections fall inside the metereological station coverage
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
