% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\name{trajreachthesea}
\alias{trajreachthesea}
\title{Check if the trajectories reach beyond the limit (the sea)}
\usage{
trajreachthesea(traj.intersections)
}
\arguments{
\item{traj.intersections}{A list made of a character vector and a list. The character vector is the path to each trajectory file while the list contains the first row in the trajectory file which lies over the sea}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{              A data.frame with one row for each file and 2 columns: The trajectories' path and a boolean indicating if they were kept
}\if{html}{\out{</div>}}
}
\description{
Check if the trajectories reach beyond the limit (the sea)
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
