% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tidyverse.R
\name{vector_to_tibble}
\alias{vector_to_tibble}
\title{Convert a vector to a tibble}
\usage{
vector_to_tibble(v)
}
\arguments{
\item{v}{a named numeric vector or a list of them.}
}
\value{
a tibble.
}
\description{
Convert a named vector (or a list of them) into a tibble, using
their names as column names.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
