% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{YEAR.SEMESTERS}
\alias{YEAR.SEMESTERS}
\title{Months grouped into semesters.}
\format{
An object of class \code{character} of length 12.
}
\usage{
YEAR.SEMESTERS
}
\description{
Months grouped into semesters.
}
\keyword{datasets}
