% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{YEAR.TRIMESTERS}
\alias{YEAR.TRIMESTERS}
\title{Months grouped into trimesters.}
\format{
An object of class \code{character} of length 12.
}
\usage{
YEAR.TRIMESTERS
}
\description{
Months grouped into trimesters.
}
\keyword{datasets}
