% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{YEAR.YEAR}
\alias{YEAR.YEAR}
\title{Months grouped into a year.}
\format{
An object of class \code{character} of length 12.
}
\usage{
YEAR.YEAR
}
\description{
Months grouped into a year.
}
\keyword{datasets}
