% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{BRIEFCASE.COLNAMES}
\alias{BRIEFCASE.COLNAMES}
\title{The name of the columns in the file resulting from processing briefcases' metadata.}
\format{
An object of class \code{character} of length 17.
}
\usage{
BRIEFCASE.COLNAMES
}
\description{
The name of the columns in the file resulting from processing briefcases' metadata.
}
\keyword{datasets}
