% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{build_grid}
\alias{build_grid}
\title{Build a grid}
\usage{
build_grid(
  origin_lon,
  origin_lat,
  min_lon,
  max_lon,
  min_lat,
  max_lat,
  grid_resolution,
  crs
)
}
\arguments{
\item{origin_lon, origin_lat}{a numeric(1). Coordinate of origin.}

\item{min_lon, max_lon}{a numeric. Grid's mininum and maximum longitude
value.}

\item{min_lat, max_lat}{a numeric. Grid's mininum and maximum latitudes
values.}

\item{grid_resolution}{A numeric. Grid's resolution.}

\item{crs}{A numeric. Coordinate reference system (EPSG).}
}
\value{
an sf object (polygon).
}
\description{
Build a grid that includes the given origin as a vertex.
}
