% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame.R
\name{cast_df_cols}
\alias{cast_df_cols}
\title{Cast data frame columns}
\usage{
cast_df_cols(.data, ctypes)
}
\arguments{
\item{.data}{a data frame.}

\item{ctypes}{a character. Column types for the given data frame.}
}
\value{
a data frame.
}
\description{
Update a data frame's column names and change their types.
}
