% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_area.R
\name{compute_frequency_grid}
\alias{compute_frequency_grid}
\title{Aggregate trajectories' vertices using a grid}
\usage{
compute_frequency_grid(
  files,
  grid_sf,
  skip = 7,
  from_row = 1,
  to_row = Inf,
  cnames = HYSPLIT.COLNAMES,
  clon = "lon",
  clat = "lat",
  cheight = "height",
  crs = 4326,
  traj_min_lon = -Inf,
  traj_max_lon = Inf,
  traj_min_lat = -Inf,
  traj_max_lat = Inf,
  traj_min_height = -Inf,
  traj_max_height = Inf,
  vert_min_lon = -Inf,
  vert_max_lon = Inf,
  vert_min_lat = -Inf,
  vert_max_lat = Inf,
  vert_min_height = -Inf,
  vert_max_height = Inf,
  min_per_vert_in_hrange = 0
)
}
\arguments{
\item{files}{a character. Paths to trajectory files (i.e. HYSPLIT files).}

\item{grid_sf}{an sf object (polygons). The grid used to aggregate
trajectories.}

\item{skip}{a numeric. Number of lines to skip from each file.}

\item{from_row, to_row}{a numeric(1). Use a subset of rows from each data
frame.}

\item{cnames}{a character. Names of the columns in each file.}

\item{clon, clat, cheight}{a character. Names of the longitude, latitude, and
height columns in \code{cnames}}

\item{crs}{a numeric. EPSG code used for the trajectories.}

\item{traj_min_lon, traj_max_lon, traj_min_lat, traj_max_lat, traj_min_height, traj_max_height}{a numeric(1). Remove trajectories which, at some vertex, fall outside of these ranges.}

\item{vert_min_lon, vert_max_lon, vert_min_lat, vert_max_lat, vert_min_height, vert_max_height}{a numeric(1). Remove vertices from trajectories falling outside of these ranges.}

\item{min_per_vert_in_hrange}{a double(1). Minimum percentage of vertices
inside height range for a trajectory to be valid.}
}
\value{
an sf object. The given grid_sf object with additional attributes ("freq".
}
\description{
Compute the number of trajectories' vertices for each cell in the grid.
}
\seealso{
\link{build_grid} for building grids.
}
