% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\docType{methods}
\name{crossdata}
\alias{crossdata}
\alias{crossdata,list,data.frame,numeric,character,numeric-method}
\alias{crossdata-generic}
\title{Get the stations' data matching the trajectory records. It takes a trajectory's first point over the sea and project its latitude to an straight line made by the corresponding metereological stations}
\usage{
crossdata(traj.intersections, stations.df, tolerance.sec, timezone,
  searchTranslation)


  \S4method{crossdata}{list,data.frame,numeric,character,numeric}(traj.intersections,
  stations.df, tolerance.sec, timezone, searchTranslation)
}
\arguments{
\item{traj.intersections}{A list made of a character vector and a list. The character vector is the path to each trajectory file while the list contains the first row in the trajectory file which lies over the sea}

\item{stations.df}{A data.frame of metereological station data. It must have at least 4 columns: name, longitude, latitude, and file c("name", "lon", "lat", "file")}

\item{tolerance.sec}{A numeric. A tolerance used when comparing dates}

\item{timezone}{A character. The time zone. i.e. "GMT"}

\item{searchTranslation}{A numeric. The number of seconds for searching metereological station data}
}
\value{
A list of numeric. Each number is the interpolation result from the matching stations
}
\description{
Get the stations' data matching the trajectory records. It takes a trajectory's first point over the sea and project its latitude to an straight line made by the corresponding metereological stations
}

