% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{df2text}
\alias{df2text}
\title{Convert a data frame to a string}
\usage{
df2text(a.df)
}
\arguments{
\item{a.df}{A data.frame}
}
\value{
A length-1 character vector
}
\description{
Convert a data frame to a length-1 character vector
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
