% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{estimate_ghg}
\alias{estimate_ghg}
\title{Estimate GHG}
\usage{
estimate_ghg(
  cross_latitude,
  cross_date,
  stations_tb,
  new_col,
  clat = "lat",
  cgas = "gas",
  cname = "name",
  cvalue = "value"
)
}
\arguments{
\item{cross_latitude}{A numeric(1). Latitude used for estimation.}

\item{cross_date}{a date(1). Date used for estimation.}

\item{stations_tb}{A tibble. Metereological stations' data.}

\item{new_col}{A character(1). Name of the new column.}

\item{clat}{A character(1). Name of the latitude column in stations_tb.}

\item{cgas}{A character(1). Name of the column with the name of the gas in
stations_tb.}

\item{cname}{A character(1). Name of the column with the station name in
stations_tb.}

\item{cvalue}{A character(1). Name of the column with the values of the
variable stored in the column cname.}
}
\value{
A named vector (double). One for each gas in the stations'
measurements.
}
\description{
Estimate the GreenHause Gass concentration at certain latitude using
metereological data from nearby stations.
}
