% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{files2df}
\alias{files2df}
\title{Read text files into data.frames}
\usage{
files2df(file.vec, header, skip, cnames)
}
\arguments{
\item{file.vec}{A vector of character. The paths to the input files}

\item{header}{A logical. Do the files have a header row?}

\item{skip}{A numeric. Lines to skip from the top of the file}

\item{cnames}{A vector of character. The column names of the data in the files}
}
\value{
A list of data.frames. The list names matches the file names
}
\description{
Read text files into data.frames (one per file)
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
