% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{filter_traj}
\alias{filter_traj}
\title{Filter trajectories}
\usage{
filter_traj(
  traj,
  clon = "lon",
  clat = "lat",
  cheight = "height",
  traj_min_lon = -Inf,
  traj_max_lon = Inf,
  traj_min_lat = -Inf,
  traj_max_lat = Inf,
  traj_min_height = -Inf,
  traj_max_height = Inf,
  min_per_vert_in_hrange = 0
)
}
\arguments{
\item{traj}{Either a data frame or a list of them. Each data frame contains
data of a single trajectory.}

\item{clon, clat, cheight}{a character(1). Names of the longitude, latitude,
and height columns in the given data frame.}

\item{traj_min_lon, traj_max_lon, traj_min_lat, traj_max_lat, traj_min_height, traj_max_height}{a numeric(1). Remove trajectories which, at some vertex, fall outside of
these ranges.}

\item{min_per_vert_in_hrange}{a double(1). Minimum percentage of vertices
inside height range for a trajectory to be valid.}
}
\value{
Either a data frame or a list ot them.
}
\description{
Filter whole trajectories from the given list of data frames.
}
