% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{filter_traj_vertices}
\alias{filter_traj_vertices}
\title{Filter trajectories' vertices}
\usage{
filter_traj_vertices(
  traj,
  clon = "lon",
  clat = "lat",
  cheight = "height",
  from_row = 1,
  to_row = Inf,
  vert_min_lon = -Inf,
  vert_max_lon = Inf,
  vert_min_lat = -Inf,
  vert_max_lat = Inf,
  vert_min_height = -Inf,
  vert_max_height = Inf
)
}
\arguments{
\item{traj}{Either a data frame or a list of them. Each data frame contains
data of a single trajectory.}

\item{clon, clat, cheight}{a character(1). Names of the longitude, latitude,
and height columns in the given data frame.}

\item{from_row, to_row}{a numeric(1). Use a subset of rows from each data
frame.}

\item{vert_min_lon, vert_max_lon, vert_min_lat, vert_max_lat, vert_min_height, vert_max_height}{a numeric(1). Remove vertices from trajectories falling outside of these
ranges.}
}
\value{
Either a data frame or a list of them.
}
\description{
Filter the trajectories' vertices in the given data frame or list of data
frames.
}
