% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\name{filterTrajHeight}
\alias{filterTrajHeight}
\title{Filter trajectories by height}
\usage{
filterTrajHeight(file.vec, above, cnames)
}
\arguments{
\item{file.vec}{A character vector. The paths to the input files}

\item{above}{A length-1 numeric. The treshold}

\item{cnames}{A character vector. The column names of the hysplit files}
}
\value{
A data frame with one row for each file and 2 columns: The trajectories' path and a boolean indicating if they meet the test
}
\description{
Check if the trajectories are above the given treshold
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
