% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\docType{methods}
\name{filterTrajHeight}
\alias{filterTrajHeight}
\alias{filterTrajHeight,character,numeric-method}
\alias{filterTrajHeight-generic}
\title{Check if the trajectories are above the given treshold}
\usage{
filterTrajHeight(file.vec, above)

\S4method{filterTrajHeight}{character,numeric}(file.vec, above)
}
\arguments{
\item{file.vec}{A character vector. The paths to the input files}

\item{above}{A numeric treshold}
}
\value{
A data.frame with one row for each file and 2 columns: The trajectories' path and a boolean indicating if they meet the test
}
\description{
Check if the trajectories are above the given treshold
}

