% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R, R/util.R
\name{get_os}
\alias{get_os}
\alias{getHistoryAC}
\alias{get_os}
\title{Get data from briefcases}
\usage{
getHistoryAC(file.in)

get_os()
}
\arguments{
\item{file.in}{A character. The path to a collect report file. i.e /home/user/PFP_3723_ALF_2010_02_17.txt}
}
\value{
A data.frame with the results of comamnds "HISTORY> A" and "HISTORY> C"

A single string
}
\description{
Extract history A and C from collected report files

Get the operative system's name. Taken from https://www.r-bloggers.com/identifying-the-os-from-r/
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}

Alber Sanchez, \email{alber.ipia@inpe.br}
}
