% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_os}
\alias{get_os}
\title{Get the operative system's name}
\usage{
get_os()
}
\value{
A single string
}
\description{
Get the operative system's name. Taken from https://www.r-bloggers.com/identifying-the-os-from-r/
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
