% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggplot2.R
\name{get_profile_plot}
\alias{get_profile_plot}
\title{Build a profile plot from the given trajectories}
\usage{
get_profile_plot(
  data_tb,
  cprofileid,
  cghg,
  cheight,
  add_mean_line = TRUE,
  mean_line_color = "black",
  mean_line_width = 2,
  add_sd_shadow = TRUE,
  xlab = "Height",
  ylab = "GHG concentration",
  title = NA
)
}
\arguments{
\item{data_tb}{A tibble with backtrajectory data.}

\item{cprofileid}{A character(1). Name of the column that identifies each
backtrajectory.}

\item{cghg}{A character(1). Name of the column with GHG concentrations.}

\item{cheight}{A character(1). Name of the column with sample heights.}

\item{add_mean_line}{A logical(1). Should be added a line representing the
profile's mean?}

\item{mean_line_color}{A character(1). Name of the color for the profile's
mean line.}

\item{mean_line_width}{A numeric(1). Width of the mean line.}

\item{add_sd_shadow}{A logical(1). Shoulb be added a shadow showing the
profile's standard deviation?}

\item{xlab, ylab}{A character(1). Label of the axes.}

\item{title}{A character(1). Plot title.}
}
\value{
A \code{ggplot2} object.
}
\description{
Build a plot object by grouping the given trajectories into profiles.
}
