% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{get_trajectory_metadata}
\alias{get_trajectory_metadata}
\title{Get metadata from trajectory file names}
\usage{
get_trajectory_metadata(
  files,
  cnames = names(TRAJECTORY.COLNAMES),
  ctypes = TRAJECTORY.COLNAMES
)
}
\arguments{
\item{files}{a character. Path to trajectory files.}

\item{cnames}{a character. Column names for the resulting data frame.}

\item{ctypes}{a character. Column data types for the resulting data frame.}
}
\value{
a data frame.
}
\description{
Build a data frame with metadata extracted from the given filenames.
}
