% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{get_trajectory_metadata}
\alias{get_trajectory_metadata}
\title{Get metadata from trajectory files}
\usage{
get_trajectory_metadata(
  files,
  cnames = TRAJECTORY.COLNAMES,
  m_period = YEAR.TRIMESTERS
)
}
\arguments{
\item{files}{a character. Path to trajectory files.}

\item{cnames}{a named character. The vector's names are the names for the
resulting data frame and its values are their data types.}

\item{m_period}{a vector with 12 elements identifying a period for each
month (e.g. trimester, semester, etc.).}
}
\value{
a data frame.
}
\description{
Build a data frame with metadata extracted from the given filenames.
}
