% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{getDataFromProfile}
\alias{getDataFromProfile}
\title{Get the data from profile name}
\usage{
getDataFromProfile(prof.name)
}
\arguments{
\item{prof.name}{A character. The name of the profile. i.e. RBA_2016_05_16}
}
\value{
A data.frame with columns c("site", "year", "month", "day")
}
\description{
Get the data from profile name
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
