% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{intersect_trajectories}
\alias{intersect_trajectories}
\title{Intersect trajectories}
\usage{
intersect_trajectories(
  traj_ls,
  limit,
  clon = "longitude",
  clat = "latitude",
  crs = 4326,
  row_after = TRUE
)
}
\arguments{
\item{traj_ls}{a list of data frames. Eacha data frame represents a
trajectory.}

\item{limit}{an sf object (line). A limit.}

\item{clon}{a character(1). The name of the longitude (x) column.}

\item{clat}{a character(1). The name of the latitude (x) column.}

\item{crs}{a numeric. The EPSG code of the trajectories.}

\item{row_after}{a logical. Should we return the id of the row after the
limit? If no, return the row id before the limit.}
}
\value{
a numeric indicating a row id for each trajectory in the input list.
}
\description{
Identify trajectories' segments that cross a limit.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
