% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\docType{methods}
\name{intersectTraj}
\alias{intersectTraj}
\alias{intersectTraj,character,SpatialLinesDataFrame-method}
\alias{intersectTraj-generic}
\title{Find trajectories' first point over the sea.}
\usage{
intersectTraj(file.vec, limit.in)

\S4method{intersectTraj}{character,SpatialLinesDataFrame}(file.vec, limit.in)
}
\arguments{
\item{file.vec}{A character vector. The paths to the input files}

\item{limit.in}{A SpatialLinesDataFrame object which is used to intersect the trajectories.}
}
\value{
A list made of a character vector and 2 lists. The character vector is the path to each trajectory file. The list contains the first row in the trajectory file which lies over the sea
}
\description{
Find trajectories' first point over the sea.
}

