% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\name{plotTrajbackground}
\alias{plotTrajbackground}
\title{Plot profiles and trajectories}
\usage{
plotTrajbackground(file.in, path.out, traj.interpol, traj.intersections,
  traj.plot, device, map.xlim, map.ylim, map.height, map.width, sec.width,
  sec.height, prof.height, prof.width, nsd, maxfm.ppm, stations.df, plot2file,
  logger, trajCnames, obsCnames, profileCnames, trajFileMet)
}
\arguments{
\item{file.in}{A character. The path to a filtered observations file or flag filtered raw data}

\item{path.out}{A character. The path to a folder to store the results}

\item{traj.interpol}{A list of numeric. The interpolated values of teh trajectories over the sea}

\item{traj.intersections}{A list made of a character vector and a list. The character vector is the path to each trajectory file while the list contains the first row in the trajectory file which lies over the sea}

\item{traj.plot}{A vector of character. The file path to trajectories to plot additional to those in traj.intersections[[1]]}

\item{device}{A character. Image format, i.e. PNG}

\item{map.xlim}{A numeric vector. Map's min & max longitude}

\item{map.ylim}{A numeric vector. Map's min & max latitude}

\item{map.height}{A numeirc. Map image size}

\item{map.width}{A numeirc. Map image size}

\item{sec.width}{A numeric. Crosssection map image size}

\item{sec.height}{A numeric. Crosssection map image size}

\item{prof.height}{A numeric. Profile image size}

\item{prof.width}{A numeric. Profile image size}

\item{nsd}{A numeric. Number of standard deviations to use to filter the interpolated data}

\item{maxfm.ppm}{A numeric. Maximum number of units away from the central tendency}

\item{stations.df}{A data.frame with metereological station data. It must contain at least the columns c("name", "lon", "lat")}

\item{plot2file}{A logical or NA. Should plots be stored as files? Use NA to not plot anything}

\item{logger}{A log4r object}

\item{trajCnames}{A character. Column names of hysplit files}

\item{obsCnames}{A character. Filtered observation column names}

\item{profileCnames}{A character. Column names of the pofiles}

\item{trajFileMet}{A character. Metadata included in the trajectory's file names}
}
\value{
A list of objects. A character vector with the paths to the plot files, and a data.frame with merged data
}
\description{
Plot the input data into map, section, and profile graphs. These plots are stored in disk
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
