% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\docType{methods}
\name{plotTrajbackground}
\alias{plotTrajbackground}
\alias{plotTrajbackground,cqmaTools-method}
\alias{plotTrajbackground-generic}
\title{Plot the input data into map, section, and profile graphs. These plots are stored in disk}
\usage{
plotTrajbackground(file.in, path.out, traj.interpol, traj.intersections, device,
  map.xlim, map.ylim, map.height, map.width, sec.width, sec.height, prof.height,
  prof.width, nsd, stations.df)

\S4method{plotTrajbackground}{cqmaTools}(file.in, path.out, traj.interpol,
  traj.intersections, device, map.xlim, map.ylim, map.height, map.width,
  sec.width, sec.height, prof.height, prof.width, nsd, stations.df)
}
\arguments{
\item{file.in}{A character. The path to a filtered observations file or flag filtered raw data}

\item{path.out}{A character. The path to a folder to store the results}

\item{traj.interpol}{A list of numeric. The interpolated values of teh trajectories over the sea}

\item{traj.intersections}{A list made of a character vector and a list. The character vector is the path to each trajectory file while the list contains the first row in the trajectory file which lies over the sea}

\item{device}{A character. Image format, i.e. PNG}

\item{map.xlim}{A numeric vector. Map's min & max longitude}

\item{map.ylim}{A numeric vector. Map's min & max latitude}

\item{map.height}{A numeirc. Map image size}

\item{map.width}{A numeirc. Map image size}

\item{sec.width}{A numeric. Crosssection map image size}

\item{sec.height}{A numeric. Crosssection map image size}

\item{prof.height}{A numeric. Profile image size}

\item{prof.width}{A numeric. Profile image size}

\item{nsd}{A numeric. Number of standard deviations to use to filter the interpolated data}

\item{stations.df}{A data.frame with metereological station data. It must contain at least the columns c("name", "lon", "lat")}
}
\value{
A character vector. The paths to the plot files
}
\description{
Plot the input data into map, section, and profile graphs. These plots are stored in disk
}

