% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\docType{methods}
\name{plotTrajYear}
\alias{plotTrajYear}
\alias{plotTrajYear,character,character,character,numeric,numeric,numeric,numeric,data.frame-method}
\alias{plotTrajYear-generic}
\title{Plot the trajectories grouped by year}
\usage{
plotTrajYear(file.vec, path.out, device, map.xlim, map.ylim, map.height,
  map.width, stations.df)


  \S4method{plotTrajYear}{character,character,character,numeric,numeric,numeric,numeric,data.frame}(file.vec,
  path.out, device, map.xlim, map.ylim, map.height, map.width, stations.df)
}
\arguments{
\item{file.vec}{A character vector. The paths to  trajectory files}

\item{path.out}{Path to the folder for storing the resulting files}

\item{device}{A character. Image format, i.e. PNG}

\item{map.xlim}{A numeric vector. Map's min & max longitude}

\item{map.ylim}{A numeric vector. Map's min & max latitude}

\item{map.height}{A numeirc. Map image size}

\item{map.width}{A numeirc. Map image size}

\item{stations.df}{A data.frame with metereological station data. It must contain at least the columns c("name", "lon", "lat")}
}
\value{
A list. The path to the created files
}
\description{
Plot the trajectories grouped by year
}

