% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence_area.R
\name{raster2convexhull}
\alias{raster2convexhull}
\title{Build a convex hull around a raster}
\usage{
raster2convexhull(r)
}
\arguments{
\item{r}{A raster (terra).}
}
\value{
A polygon (sf).
}
\description{
Build a convex hull round the pixels in the given raster which
are not NULL.
}
\author{
Alber Sánchez, \email{alber.ipia@inpe.br}
}
