% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{removeHeader}
\alias{removeHeader}
\title{Read a file and remove its header}
\usage{
removeHeader(file_path, col_names, skip)
}
\arguments{
\item{file_path}{A length-1 character vector. A path to a file}

\item{col_names}{A character. The column names fo the returned data frame}

\item{skip}{A length-1 numeric. Number of lines to skip when reading the file}
}
\value{
A data.frame
}
\description{
Read a file and remove its header
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
