% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\docType{methods}
\name{removeHeaders}
\alias{removeHeaders}
\alias{removeHeaders,character,character,numeric-method}
\alias{removeHeaders-generic}
\title{Remove headers from hysplit files}
\usage{
removeHeaders(file.vec, path.out, skip)

\S4method{removeHeaders}{character,character,numeric}(file.vec, path.out, skip)
}
\arguments{
\item{file.vec}{A character vector. The paths to the input files}

\item{path.out}{Path to the folder for storing the resulting files}

\item{skip}{Numeric. Number of lines to remove from the beginning of the file}
}
\value{
A list of the created file paths
}
\description{
Remove headers from hysplit files
}

