% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{removeHeaders}
\alias{removeHeaders}
\title{Remove header from files}
\usage{
removeHeaders(file.vec, path.out, skip, cnames)
}
\arguments{
\item{file.vec}{A character vector. The paths to the input files}

\item{path.out}{A length-1 character. The path to the folder for storing the resulting files}

\item{skip}{A length-1 numeric. Number of lines to remove from the beginning of each file}

\item{cnames}{A character. The names of the columns of the returned data frame}
}
\value{
A list of paths to the created files
}
\description{
Remove header from files
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
