% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{smooth_neigh_na}
\alias{smooth_neigh_na}
\title{Smooth raster using NA neighbors}
\usage{
smooth_neigh_na(r, w = 3, threshold = 7)
}
\arguments{
\item{r}{a raster (terra).}

\item{w}{an integer. The window size.}

\item{threshold}{an integer.}
}
\value{
A raster (terra).
}
\description{
Transform the cells into NA depending on the number of NA neighbors.
}
