% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\docType{methods}
\name{splitRawdata}
\alias{splitRawdata}
\alias{splitRawdata,character,character,character,character-method}
\alias{splitRawdata-generic}
\title{Split a single raw data file. The output files are prefixed with "traj_specs_".}
\usage{
splitRawdata(file.in, path.out, colname, keepFlags)

\S4method{splitRawdata}{character,character,character,character}(file.in,
  path.out, colname, keepFlags)
}
\arguments{
\item{file.in}{Path to a raw data file}

\item{path.out}{Path to the folder for storing the resulting files}

\item{colname}{A character. The name of a column where the flags are located}

\item{keepFlags}{A character vector. Flags to keep in the raw data. i.e. c("...", "..>")}
}
\value{
A character. The name of the new file
}
\description{
Split a single raw data file. The output files are prefixed with "traj_specs_".
}

