% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states_sf}
\alias{states_sf}
\title{State polygons}
\format{
\subsection{\code{states_sf}}{

A sf object with 4596 rows and 2 columns:
\describe{
\item{name}{State name}
\item{geometry}{Simple feature geometry list column (sfc)}
}
}
}
\source{
\url{https://www.naturalearthdata.com}
}
\usage{
states_sf
}
\description{
Administrative level 1 polygons obtained from Natural Earth. These polygons #' are used only as plot background.
}
\keyword{datasets}
