% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{traj2lines}
\alias{traj2lines}
\title{Build trajectory lines}
\usage{
traj2lines(traj_df, clon, clat)
}
\arguments{
\item{traj_df}{A data.frame with trajectory data in HYSPLIT format.}

\item{clon}{a character(1). The name of the longitude (x) column.}

\item{clat}{a character(1). The name of the latitude (x) column.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    A list of [sf::st_linestring] objects.
}\if{html}{\out{</div>}}
}
\description{
Build a line (st_linestring) for each pair of points in the
given trajectory.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
