% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqmaTools.R
\docType{methods}
\name{trajreachthesea}
\alias{trajreachthesea}
\alias{trajreachthesea,list-method}
\alias{trajreachthesea-generic}
\title{Check if the trajectories reach the sea}
\usage{
trajreachthesea(traj.intersections)

\S4method{trajreachthesea}{list}(traj.intersections)
}
\arguments{
\item{traj.intersections}{A list made of a character vector and a list. The character vector is the path to each trajectory file while the list contains the first row in the trajectory file which lies over the sea}
}
\value{
A data.frame with one row for each file and 2 columns: The trajectories' path and a boolean indicating if they were kept
}
\description{
Check if the trajectories reach the sea
}

