% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{ydec2date}
\alias{ydec2date}
\title{Transform a decimal date to a date-format string}
\usage{
ydec2date(ydec)
}
\arguments{
\item{ydec}{a numeric(1) representing a date as a decimal year. e.g.
2000.0013661202.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{    A character. e.g. 2000-01-01 11:59:59.999410
}\if{html}{\out{</div>}}
}
\description{
Transform a decimal-year date into a date-format string
}
